Support browser path false of requires without ext
It's pretty common for node modules to use require() without specifying the file extension, so adding this will most likely make a few more libraries work with rollup. One project doing this is for instance `object-assign` which doesn't work currently. https://github.com/substack/object-inspect/issues/18

Considering the line that checks for changed paths below does exactly the same it's probably just an oversight that this was missing.